docs/requirements.txt
cmocean==4.0.3
colorcet==3.1.0
enum-tools==0.12.0
imageio==2.34.2
imageio-ffmpeg==0.5.1
jupyter_sphinx==0.5.3
jupyterlab==4.2.3
lxml==5.2.2
meshio==5.3.5
mypy==1.10.1
mypy-extensions==1.0.0
numpydoc==1.7.0
osmnx==1.9.3
pydata-sphinx-theme==0.15.4
pypandoc==1.13
pytest-sphinx==0.6.3
scipy==1.14.0
sphinx==7.3.7
sphinx-autobuild==2024.4.16
sphinx-copybutton==0.5.2
sphinx-design==0.6.0
sphinx-gallery==0.16.0
sphinx-notfound-page==1.0.2
sphinx-toolbox==3.6.0
sphinxcontrib-asciinema==0.4.2
sphinxcontrib-websupport==1.2.7
sympy==1.12.1
trame==3.6.3
trame-vtk==2.8.9
trame-vuetify==2.6.0
trimesh==4.4.1
sphinx_rtd_theme==1.3.0rc1
accessible-pygments==0.0.5
beautifulsoup4==4.12.3
pydata-sphinx-theme==0.15.4
soupsieve==2.5
sphinx-math-dollar==1.2.1
trame-client==3.2.1
typing-extensions==4.12.2
phasefieldx==0.0.1
